Fix label indentation in surface blender
