+++ src/LingoEngine.SDL2/FilmLoops/SdlMemberFilmLoop.cs
@@ -1,4 +1,3 @@
-using System.Numerics;
 using System.Runtime.InteropServices;
 using AbstUI.Primitives;
 using AbstUI.SDL2.Bitmaps;
@@ -10,7 +9,6 @@ using LingoEngine.Primitives;
 using LingoEngine.SDL2.Bitmaps;
 using LingoEngine.SDL2.Inputs;
 using LingoEngine.Sprites;
-using LingoEngine.Tools;
 
 namespace LingoEngine.SDL2.FilmLoops;
 
@@ -188,7 +186,7 @@ public class SdlMemberFilmLoop : ILingoFrameworkMemberFilmLoop, IDisposable
                 freeSrc = true;
             }
 
-            BlendSurface(surface, srcSurf, info.Transform.Matrix, info.Alpha);
+            SdlSurfaceBlender.BlendSurface(surface, srcSurf, info.Transform.Matrix, info.Alpha);
 
             if (freeSrc)
                 SDL.SDL_FreeSurface(srcSurf);
@@ -203,91 +201,6 @@ public class SdlMemberFilmLoop : ILingoFrameworkMemberFilmLoop, IDisposable
 #if DEBUG
     public void DebugToDisk(nint texture, string filName) => SdlTexture2D.DebugToDisk(_sdlRootContext.Renderer, texture, filName);
 #endif
-    /// <summary>
-    /// Blends <paramref name="src"/> onto <paramref name="dest"/> using an affine transform.
-    /// </summary>
-    private static unsafe void BlendSurface(nint dest, nint src, Matrix3x2 transform, float _)
-    {
-        SDL.SDL_LockSurface(dest);
-        SDL.SDL_LockSurface(src);
-
-        var dSurf = Marshal.PtrToStructure<SDL.SDL_Surface>(dest);
-        var sSurf = Marshal.PtrToStructure<SDL.SDL_Surface>(src);
-        var dFmt = Marshal.PtrToStructure<SDL.SDL_PixelFormat>(dSurf.format);
-        var sFmt = Marshal.PtrToStructure<SDL.SDL_PixelFormat>(sSurf.format);
-
-        if (dSurf.format == nint.Zero || sSurf.format == nint.Zero) goto UNLOCK;
-        if (dFmt.BytesPerPixel != 4 || sFmt.BytesPerPixel != 4) goto UNLOCK;
-
-        static void Unpack(uint px, in SDL.SDL_PixelFormat f, out byte r, out byte g, out byte b, out byte a)
-        {
-            r = (byte)((px & f.Rmask) >> f.Rshift);
-            g = (byte)((px & f.Gmask) >> f.Gshift);
-            b = (byte)((px & f.Bmask) >> f.Bshift);
-            a = (byte)((px & f.Amask) >> f.Ashift);
-            // normalize to 8-bit if losses exist
-            r = (byte)(r * 255 / (f.Rmask >> f.Rshift));
-            g = (byte)(g * 255 / (f.Gmask >> f.Gshift));
-            b = (byte)(b * 255 / (f.Bmask >> f.Bshift));
-            a = (byte)(a * 255 / (f.Amask >> f.Ashift));
-        }
-        static uint Pack(byte r, byte g, byte b, byte a, in SDL.SDL_PixelFormat f)
-        {
-            uint R = r * (f.Rmask >> f.Rshift) / 255 << f.Rshift;
-            uint G = g * (f.Gmask >> f.Gshift) / 255 << f.Gshift;
-            uint B = b * (f.Bmask >> f.Bshift) / 255 << f.Bshift;
-            uint A = a * (f.Amask >> f.Ashift) / 255 << f.Ashift;
-            return R & f.Rmask | G & f.Gmask | B & f.Bmask | A & f.Amask;
-        }
-
-        Matrix3x2.Invert(transform, out var inv);
-
-        Vector2[] pts = {
-        Vector2.Transform(Vector2.Zero, transform),
-        Vector2.Transform(new Vector2(sSurf.w, 0), transform),
-        Vector2.Transform(new Vector2(sSurf.w, sSurf.h), transform),
-        Vector2.Transform(new Vector2(0, sSurf.h), transform)
-    };
-        int minX = (int)MathF.Floor(pts.Min(p => p.X));
-        int maxX = (int)MathF.Ceiling(pts.Max(p => p.X));
-        int minY = (int)MathF.Floor(pts.Min(p => p.Y));
-        int maxY = (int)MathF.Ceiling(pts.Max(p => p.Y));
-
-        byte* dpix = (byte*)dSurf.pixels;
-        byte* spix = (byte*)sSurf.pixels;
-
-        int totalPixels = (maxX - minX) * (maxY - minY);
-        ParallelHelper.For(minY, maxY, totalPixels, y =>
-        {
-            if ((uint)y >= (uint)dSurf.h) return;
-            byte* drow = dpix + y * dSurf.pitch;
-
-            for (int x = minX; x < maxX; x++)
-            {
-                if ((uint)x >= (uint)dSurf.w) continue;
-
-                var srcPos = Vector2.Transform(new Vector2(x + 0.5f, y + 0.5f), inv);
-                int sx = (int)MathF.Floor(srcPos.X);
-                int sy = (int)MathF.Floor(srcPos.Y);
-                if ((uint)sx >= (uint)sSurf.w || (uint)sy >= (uint)sSurf.h) continue;
-
-                uint* sp = (uint*)(spix + sy * sSurf.pitch + sx * 4);
-                byte r, g, b, a;
-                Unpack(*sp, sFmt, out r, out g, out b, out a);
-                if (a == 0) continue; // skip fully transparent
-
-                uint* dp = (uint*)(drow + x * 4);
-                *dp = Pack(r, g, b, a, dFmt); // no extra blending
-            }
-        });
-
-    UNLOCK:
-        SDL.SDL_UnlockSurface(src);
-        SDL.SDL_UnlockSurface(dest);
-    }
-
-
-
     public void Dispose() { Unload(); }
